requirements.txt
Django==3.0.3
ipython==7.12.0
django-durationwidget==1.0.5
djangorestframework==3.11.0
Markdown==3.2
django-filter==2.2.0


